+++ docs/grpc/docs.md
@@ -3025,6 +3025,7 @@ Additionally, the first and last points of each GeoLineString must be the same.
 | search_params | [SearchParams](#qdrant-SearchParams) | optional | Search params for when there is no prefetch. |
 | score_threshold | [float](#float) | optional | Return points with scores better than this threshold. |
 | limit | [uint64](#uint64) | optional | Max number of points. Default is 10 |
+| lookup_from | [LookupLocation](#qdrant-LookupLocation) | optional | Name of the collection to use for points lookup, if not specified - use current collection |
 
 
 
@@ -3093,7 +3094,7 @@ For example, if `oversampling` is 2.4 and `limit` is 100, then 240 vectors will 
 | with_payload | [WithPayloadSelector](#qdrant-WithPayloadSelector) | optional | Options for specifying which payload to include or not. |
 | read_consistency | [ReadConsistency](#qdrant-ReadConsistency) | optional | Options for specifying read consistency guarantees. |
 | shard_key_selector | [ShardKeySelector](#qdrant-ShardKeySelector) | optional | Specify in which shards to look for the points, if not specified - look in all shards. |
-| with_lookup | [WithLookup](#qdrant-WithLookup) | optional | Options for specifying how to use the group id to lookup points in another collection |
+| lookup_from | [LookupLocation](#qdrant-LookupLocation) | optional | Name of the collection to use for points lookup, if not specified - use current collection |
