+++ docs/redoc/master/openapi.json
@@ -11417,11 +11417,12 @@
               }
             ]
           },
-          "with_lookup": {
-            "description": "Options for specifying where to lookup vector ids from during recommendation",
+          "lookup_from": {
+            "description": "The location used to lookup vectors. If not specified - use current collection. Note: the other collection should have the same vector size as the current collection",
+            "default": null,
             "anyOf": [
               {
-                "$ref": "#/components/schemas/WithLookupInterface"
+                "$ref": "#/components/schemas/LookupLocation"
               },
               {
                 "nullable": true
@@ -11501,6 +11502,18 @@
             "format": "uint",
             "minimum": 0,
             "nullable": true
+          },
+          "lookup_from": {
+            "description": "The location used to lookup vectors. If not specified - use current collection. Note: the other collection should have the same vector size as the current collection",
+            "default": null,
+            "anyOf": [
+              {
+                "$ref": "#/components/schemas/LookupLocation"
+              },
+              {
+                "nullable": true
+              }
+            ]
           }
         }
       },
